Remove `include/grpc/impl/codegen/sync.h`
