+++ include/grpc/impl/codegen/sync_generic.h
@@ -21,31 +21,9 @@
 
 // IWYU pragma: private, include <grpc/support/sync.h>
 
-/* Generic type definitions for gpr_sync. */
+#include <grpc/support/port_platform.h>
 
-#include <grpc/impl/codegen/port_platform.h>
-
-#include <grpc/impl/codegen/atm.h>
-
-/* gpr_event */
-typedef struct {
-  gpr_atm state;
-} gpr_event;
-
-#define GPR_EVENT_INIT \
-  { 0 }
-
-/* gpr_refcount */
-typedef struct {
-  gpr_atm count;
-} gpr_refcount;
-
-/* gpr_stats_counter */
-typedef struct {
-  gpr_atm value;
-} gpr_stats_counter;
-
-#define GPR_STATS_INIT \
-  { 0 }
+/// TODO(chengyuc): Remove this file after solving compatibility.
+#include <grpc/support/sync_generic.h>
 
 #endif /* GRPC_IMPL_CODEGEN_SYNC_GENERIC_H */
