+++ include/grpc/impl/codegen/sync_windows.h
@@ -21,22 +21,9 @@
 
 // IWYU pragma: private, include <grpc/support/sync.h>
 
-#include <grpc/impl/codegen/port_platform.h>
+#include <grpc/support/port_platform.h>
 
-#ifdef GPR_WINDOWS
-
-#include <grpc/impl/codegen/sync_generic.h>
-
-typedef struct {
-  CRITICAL_SECTION cs; /* Not an SRWLock until Vista is unsupported */
-  int locked;
-} gpr_mu;
-
-typedef CONDITION_VARIABLE gpr_cv;
-
-typedef INIT_ONCE gpr_once;
-#define GPR_ONCE_INIT INIT_ONCE_STATIC_INIT
-
-#endif /* GPR_WINDOWS */
+/// TODO(chengyuc): Remove this file after solving compatibility.
+#include <grpc/support/sync_windows.h>
 
 #endif /* GRPC_IMPL_CODEGEN_SYNC_WINDOWS_H */
