+++ include/grpc/support/sync_windows.h
@@ -21,6 +21,20 @@
 
 #include <grpc/support/port_platform.h>
 
-#include <grpc/impl/codegen/sync_windows.h>  // IWYU pragma: export
+#ifdef GPR_WINDOWS
+
+#include <grpc/support/sync_generic.h>
+
+typedef struct {
+  CRITICAL_SECTION cs; /* Not an SRWLock until Vista is unsupported */
+  int locked;
+} gpr_mu;
+
+typedef CONDITION_VARIABLE gpr_cv;
+
+typedef INIT_ONCE gpr_once;
+#define GPR_ONCE_INIT INIT_ONCE_STATIC_INIT
+
+#endif /* GPR_WINDOWS */
 
 #endif /* GRPC_SUPPORT_SYNC_WINDOWS_H */
